Port the full sim step: byte-exact full-day event stream
Complete the deterministic simulation port. sim/world.h + sim/step.c
mirror state.ts + step.ts tick-for-tick: herder modes, sheep tempers
and wandering, mishaps, weather, streaks, nemesis, jailbreaks, lunch
and the lunch thief, the rival, the dog's moments, milestones, the
reading/book system, progression, frustration, and the wall-clock
governor. sim/book.{c,h} carry the book catalogue.

Fix: HerderSheep positions (x/y/tx/ty/home) were int and truncated all
fractional movement, so sheep never reached wander targets and the
herder could never close the last tile to catch them. Now double, as
in the web float64 model.

Golden extended with a full-day event transcript (EV/EVEND) for seeds
'seed' and 'grudge'; host test asserts all 626 events tick-exact.

Host tests: 1605 checks, 0 failures. All modules compile for SH-4.

+++ src/core/sim/flock.c
@@ -152,9 +152,9 @@ int herder_place_libraries(const HerderMap *map, const HerderSheep *sheep, int s
     /* byDistance: stable sort of sheep indices by pen distance (ties keep order). */
     int order[128]; for (int i = 0; i < sheepN; i++) order[i] = i;
     for (int i = 1; i < sheepN; i++) { /* insertion sort = stable */
-        int key = order[i]; double kd = map->pen_distance[sheep[key].y * n + sheep[key].x];
+        int key = order[i]; double kd = map->pen_distance[(int)sheep[key].y * n + (int)sheep[key].x];
         int j = i - 1;
-        while (j >= 0) { double jd = map->pen_distance[sheep[order[j]].y * n + sheep[order[j]].x]; if (jd <= kd) break; order[j+1] = order[j]; j--; }
+        while (j >= 0) { double jd = map->pen_distance[(int)sheep[order[j]].y * n + (int)sheep[order[j]].x]; if (jd <= kd) break; order[j+1] = order[j]; j--; }
         order[j+1] = key;
     }
     int count = LIBRARY_COUNT < sheepN ? LIBRARY_COUNT : sheepN;
